+++ ReleaseNotes.md
@@ -6,4 +6,6 @@
 
 ## Nightly
 - Inference for the Gumbel distribution
+- Replace abstract type EVA by AbstractExtremeValueModel
+- Replace abstract type fittedEVA by AbstractFittedExtremeValueModel